+++ apps/bot/handlers/top.py
@@ -8,6 +8,7 @@ from apps.bot.logger import logger
 from apps.bot.utils import update_or_create_user
 from apps.bot.utils.language import with_language, set_language_code
 from apps.bot.utils.send_cinema import send_cinema
+from apps.bot.utils.subscription import check_subscribe
 
 
 @with_language
@@ -21,30 +22,35 @@ def callback_handler_top(call: CallbackQuery, bot: TeleBot):
     )
     activate(set_language_code(call.from_user.id))
     logger.info(f"User {call.from_user.id} selected a top.")
-    cinemas = Cinema.objects.filter(is_active=True).order_by("-view_count")[:10]
-    markup = InlineKeyboardMarkup(row_width=5)
-    indexes = []
-    text = _("Top 10 cinemas:\n\n")
-    for index, cinema in enumerate(cinemas):
-        text += f"\n<b>{index + 1}.</b> {cinema.title} | 👁{cinema.view_count}"
-        indexes.append(index + 1)
-    buttons = [
-        [
-            InlineKeyboardButton(
-                f"🎬 {index}", callback_data=f"cinema_{cinemas[index - 1].id}"
-            )
-            for index in indexes
+    if check_subscribe(
+        bot=bot,
+        user_id=call.from_user.id,
+        call=call,
+    ):
+        cinemas = Cinema.objects.filter(is_active=True).order_by("-view_count")[:10]
+        markup = InlineKeyboardMarkup(row_width=5)
+        indexes = []
+        text = _("Top 10 cinemas:\n\n")
+        for index, cinema in enumerate(cinemas):
+            text += f"\n<b>{index + 1}.</b> {cinema.title} | 👁{cinema.view_count}"
+            indexes.append(index + 1)
+        buttons = [
+            [
+                InlineKeyboardButton(
+                    f"🎬 {index}", callback_data=f"cinema_{cinemas[index - 1].id}"
+                )
+                for index in indexes
+            ]
         ]
-    ]
-    markup.add(*[button for row in buttons for button in row])
-    markup.add(InlineKeyboardButton(_("◀️Back"), callback_data="back"))
-    bot.edit_message_text(
-        chat_id=call.message.chat.id,
-        message_id=call.message.message_id,
-        text=text,
-        parse_mode="HTML",
-        reply_markup=markup,
-    )
+        markup.add(*[button for row in buttons for button in row])
+        markup.add(InlineKeyboardButton(_("◀️Back"), callback_data="back"))
+        bot.edit_message_text(
+            chat_id=call.message.chat.id,
+            message_id=call.message.message_id,
+            text=text,
+            parse_mode="HTML",
+            reply_markup=markup,
+        )
 
 
 @with_language
@@ -57,29 +63,34 @@ def callback_handler_top_cinema(call: CallbackQuery, bot: TeleBot):
         is_active=True,
     )
     logger.info(f"User {call.from_user.id} selected a top.")
+    if check_subscribe(
+        bot=bot,
+        user_id=call.from_user.id,
+        call=call,
+    ):
 
-    cinema_id = call.data.split("_")[1]
-    try:
-        bot.delete_message(call.message.chat.id, call.message.message_id)
+        cinema_id = call.data.split("_")[1]
         try:
-            cinema = Cinema.objects.get(id=cinema_id)
-        except Cinema.DoesNotExist:
+            bot.delete_message(call.message.chat.id, call.message.message_id)
+            try:
+                cinema = Cinema.objects.get(id=cinema_id)
+            except Cinema.DoesNotExist:
+                bot.send_message(
+                    call.message.chat.id,
+                    _("Sorry, the cinema you are looking for does not exist."),
+                )
+                return
+            user = BotUser.objects.filter(telegram_id=call.from_user.id).first()
+            send_cinema(
+                cinema_id=cinema.id,
+                chat_id=call.message.chat.id,
+                bot=bot,
+                user=user,
+            )
+
+        except Exception as e:
+            logger.exception(f"Error while processing random request {e}")
             bot.send_message(
                 call.message.chat.id,
-                _("Sorry, the cinema you are looking for does not exist."),
+                _("Sorry, there was an error while processing your request."),
             )
-            return
-        user = BotUser.objects.filter(telegram_id=call.from_user.id).first()
-        send_cinema(
-            cinema_id=cinema.id,
-            chat_id=call.message.chat.id,
-            bot=bot,
-            user=user,
-        )
-
-    except Exception as e:
-        logger.exception(f"Error while processing random request {e}")
-        bot.send_message(
-            call.message.chat.id,
-            _("Sorry, there was an error while processing your request."),
-        )
